Refactor breadcrumb header and move to separate component for better organization
