+++ apps/nextjs/src/app/dashboard/breadcrumbs.tsx
@@ -1,4 +1,4 @@
-import { Fragment, Suspense } from "react";
+import { Fragment } from "react";
 
 import {
   Breadcrumb,
@@ -8,33 +8,6 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from "@battle-stadium/ui/breadcrumb";
-import { Separator } from "@battle-stadium/ui/separator";
-import { SidebarTrigger } from "@battle-stadium/ui/sidebar";
-
-const LAYOUT_CONSTANTS = {
-  HEADER_HEIGHT: "h-16",
-  HEADER_HEIGHT_COLLAPSED: "h-12",
-  SEPARATOR_HEIGHT: "h-4",
-} as const;
-
-export default function DashboardBreadcrumbHeader() {
-  return (
-    <header
-      className={`flex ${LAYOUT_CONSTANTS.HEADER_HEIGHT} shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[data-collapsible=icon]/sidebar-wrapper:${LAYOUT_CONSTANTS.HEADER_HEIGHT_COLLAPSED} mr-4 mt-4 rounded-xl border-none bg-neutral-900/50`}
-    >
-      <div className="flex items-center gap-2 px-4">
-        <SidebarTrigger className="ml-1" aria-label="Open Sidebar" />
-        <Separator
-          orientation="vertical"
-          className={`mr-2 ${LAYOUT_CONSTANTS.SEPARATOR_HEIGHT}`}
-        />
-        <Suspense fallback={null}>
-          <Breadcrumbs />
-        </Suspense>
-      </div>
-    </header>
-  );
-}
 
 interface BreadcrumbsProps {
   items?: {
@@ -48,7 +21,7 @@ const defaultItems = [
   { label: "Dashboard", href: "/dashboard" },
 ];
 
-function Breadcrumbs({ items = defaultItems }: BreadcrumbsProps) {
+export function Breadcrumbs({ items = defaultItems }: BreadcrumbsProps) {
   return (
     <Breadcrumb aria-label="Navigation breadcrumbs">
       <BreadcrumbList>
