+++ apps/nextjs/src/app/dashboard/layout.tsx
@@ -1,9 +1,11 @@
-import { SidebarInset, SidebarProvider } from "@battle-stadium/ui/sidebar";
+import { SidebarInset, SidebarProvider, SidebarTrigger } from "@battle-stadium/ui/sidebar";
 
 import type { ChildrenProps } from "~/types";
-import DashboardHeader from "~/app/dashboard/breadcrumb-header";
+import { Breadcrumbs } from "~/app/dashboard/breadcrumbs";
 import { BreadCrumbsProvider } from "~/components/breadcrumbs/context";
 import SideBar from "./@sidebar/sidebar";
+import { Separator } from "@battle-stadium/ui";
+import { Suspense } from "react";
 
 interface DashboardLayoutProps extends ChildrenProps {
   sidebar: React.ReactNode;
@@ -19,7 +21,7 @@ export default function DashboardLayout({
         <SideBar>{sidebar}</SideBar>
 
         <SidebarInset className="border-none bg-transparent">
-          <DashboardHeader />
+          <DashboardBreadcrumbHeader />
 
           <div
             aria-label="dashboard content"
@@ -30,5 +32,30 @@ export default function DashboardLayout({
         </SidebarInset>
       </SidebarProvider>
     </BreadCrumbsProvider>
+  );
+}
+
+const LAYOUT_CONSTANTS = {
+  HEADER_HEIGHT: "h-16",
+  HEADER_HEIGHT_COLLAPSED: "h-12",
+  SEPARATOR_HEIGHT: "h-4",
+} as const;
+
+function DashboardBreadcrumbHeader () {
+  return (
+    <header
+      className={ `flex ${LAYOUT_CONSTANTS.HEADER_HEIGHT} shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[data-collapsible=icon]/sidebar-wrapper:${LAYOUT_CONSTANTS.HEADER_HEIGHT_COLLAPSED} mr-4 mt-4 rounded-xl border-none bg-neutral-900/50` }
+    >
+      <div className="flex items-center gap-2 px-4">
+        <SidebarTrigger className="ml-1" aria-label="Open Sidebar" />
+        <Separator
+          orientation="vertical"
+          className={ `mr-2 ${LAYOUT_CONSTANTS.SEPARATOR_HEIGHT}` }
+        />
+        <Suspense fallback={ null }>
+          <Breadcrumbs />
+        </Suspense>
+      </div>
+    </header>
   );
 }
