+++ apps/nextjs/src/app/dashboard/organizations/page.tsx
@@ -1,5 +1,7 @@
 import ComingSoon from "~/components/coming-soon";
 
 export default function OrganizationsDashboard() {
-  return <ComingSoon title="Organizations Dashboard Home" />;
+  return (
+    <ComingSoon title="Organizations Dashboard Home" />
+  );
 }
